+++ internal/ui/assets_pane.go
@@ -39,7 +39,7 @@ func (p *assetsPane) build() fyne.CanvasObject {
 
 	p.list = widget.NewList(
 		func() int { return len(p.items) },
-		func() fyne.CanvasObject { return widget.NewLabel("template") },
+		func() fyne.CanvasObject { return monoLabel("template") },
 		func(i widget.ListItemID, o fyne.CanvasObject) {
 			o.(*widget.Label).SetText(assetRow(p.items[i]))
 		},